Add/update GPL license boilerplate to all code files
