+++ tests/pywndcharm_tests/test_SingleSamplePrediction.py
@@ -1,24 +1,27 @@
-#!/usr/bin/env python
-# -*- coding: utf-8 -*-
+"""
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+                                                                               
+ Copyright (C) 2015 National Institutes of Health 
 
-#
-# Copyright (C) 2014 National Institutes of Health
-# All rights reserved.
-#
-# This program is free software; you can redistribute it and/or modify
-# it under the terms of the GNU General Public License as published by
-# the Free Software Foundation; either version 2 of the License, or
-# (at your option) any later version.
-#
-# This program is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-# GNU General Public License for more details.
-#
-# You should have received a copy of the GNU General Public License along
-# with this program; if not, write to the Free Software Foundation, Inc.,
-# 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
-#
+    This library is free software; you can redistribute it and/or              
+    modify it under the terms of the GNU Lesser General Public                 
+    License as published by the Free Software Foundation; either               
+    version 2.1 of the License, or (at your option) any later version.         
+                                                                               
+    This library is distributed in the hope that it will be useful,            
+    but WITHOUT ANY WARRANTY; without even the implied warranty of             
+    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU          
+    Lesser General Public License for more details.                            
+                                                                               
+    You should have received a copy of the GNU Lesser General Public           
+    License along with this library; if not, write to the Free Software        
+    Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA  
+                                                                               
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+                                                                               
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+ Written by:  Christopher Coletta (github.com/colettace)
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~"""
 
 import sys
 if sys.version_info < (2, 7):
