+++ wndcharm/__init__.py
@@ -1,3 +1,28 @@
+"""
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+                                                                               
+ Copyright (C) 2015 National Institutes of Health 
+
+    This library is free software; you can redistribute it and/or              
+    modify it under the terms of the GNU Lesser General Public                 
+    License as published by the Free Software Foundation; either               
+    version 2.1 of the License, or (at your option) any later version.         
+                                                                               
+    This library is distributed in the hope that it will be useful,            
+    but WITHOUT ANY WARRANTY; without even the implied warranty of             
+    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU          
+    Lesser General Public License for more details.                            
+                                                                               
+    You should have received a copy of the GNU Lesser General Public           
+    License along with this library; if not, write to the Free Software        
+    Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA  
+                                                                               
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+                                                                               
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+ Written by:  Christopher Coletta (github.com/colettace)
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~"""
+
 __version__ = "unknown"
 
 from wndcharm import *
